+++ djmoney/forms/fields.py
@@ -6,7 +6,7 @@ from warnings import warn
 from django.core.exceptions import ValidationError
 from django.forms import ChoiceField, DecimalField, MultiValueField
 
-from moneyed.classes import Money
+from djmoney.money import Money
 
 from ..settings import CURRENCY_CHOICES
 from .widgets import MoneyWidget