+++ tests/conftest.py
@@ -5,7 +5,7 @@ from django import VERSION
 
 import pytest
 
-from moneyed import Money
+from djmoney.money import Money
 from tests.testapp.models import InheritorModel, ModelWithDefaultAsInt
 
 from ._compat import patch