+++ tests/test_admin.py
@@ -6,7 +6,7 @@ from django import VERSION
 
 import pytest
 
-from moneyed import Money
+from djmoney.money import Money
 
 from .testapp.models import ModelWithVanillaMoneyField
 
@@ -24,16 +24,17 @@ def get_args(value, field):
     return value, field, ''
 
 
-@pytest.mark.parametrize('value, kwargs, expected', (
-    (Money(10, 'RUB'), {}, '10.00 руб.'),  # Issue 232
-    (Money(1234), {'USE_L10N': True, 'USE_THOUSAND_SEPARATOR': True}, '1,234.00 XYZ'),  # Issue 220
-    (Money(1000, 'SAR'), {'USE_I18N': True, 'LANGUAGE_CODE': 'en-us'}, 'ر.س1,000.00'),  # Issue 196
-    (Money(1000, 'PLN'), {}, '1,000.00 zł'),  # Issue 102
-    (Money('3.33', 'EUR'), {'USE_I18N': True, 'LANGUAGE_CODE': 'de-de'}, '3.33 €'),  # Issue 90
+@pytest.mark.parametrize('value, expected', (
+    (Money(10, 'RUB'), '10.00 руб.'),  # Issue 232
+    (Money(1234), '1,234.00 XYZ'),  # Issue 220
+    (Money(1000, 'SAR'), 'ر.س1,000.00'),  # Issue 196
+    (Money(1000, 'PLN'), '1,000.00 zł'),  # Issue 102
+    (Money('3.33', 'EUR'), '3.33 €'),  # Issue 90
 ))
-def test_display_for_field(settings, value, kwargs, expected):
-    for k, v in kwargs.items():
-        setattr(settings, k, v)
+def test_display_for_field(settings, value, expected):
+    settings.USE_L10N = True
+    # This locale has no definitions in py-moneyed, so it will work for localized money representation.
+    settings.LANGUAGE_CODE = 'cs'
     assert admin_utils.display_for_field(*get_args(value, MONEY_FIELD)) == expected