+++ tests/test_managers.py
@@ -4,8 +4,9 @@ from django.db.models import F, Q
 import pytest
 
 from djmoney.models.managers import _expand_money_args, _expand_money_kwargs
+from djmoney.money import Money
 from djmoney.utils import get_amount
-from moneyed import Money
+from moneyed import Money as OldMoney
 
 from .testapp.models import ModelWithNonMoneyField
 
@@ -144,6 +145,7 @@ class TestKwargsExpand:
             (
                 ({'money': 100, 'desc': 'test'}, {'money': 100, 'desc': 'test'}),
                 ({'money': Money(100, 'USD')}, {'money': 100, 'money_currency': 'USD'}),
+                ({'money': OldMoney(100, 'USD')}, {'money': 100, 'money_currency': 'USD'}),
                 ({'money': Money(100, 'USD'), 'desc': 'test'}, {'money': 100, 'money_currency': 'USD', 'desc': 'test'}),
             )
         )